Refactor windows trap handling to look like Unix
Shuffle some files around to be more amenable to #[cfg]
